requirements.txt
fastapi
uvicorn
smtplib
python-dotenv
